refactor: 게시판 헤더 컴포넌트 분리 (#61)
- 프롭스 정리
- 프롭스 드릴링만 봐도 nuqs가 좋아보임
