+++ src/features/community-filter/ui/CommunityBoardFilters.tsx
@@ -2,62 +2,27 @@
 // TODO: nuqs 를 쓰게 되면 다시 클라이언트 컴포넌트로
 
 import Link from 'next/link'
-import { cn } from '@/shared/lib/cn'
-import { Input } from '@/shared/ui/input'
 import { Plus } from 'lucide-react'
+import CommunityCategoryTabs from '@/features/community-filter/ui/CommunityCategoryTabs'
+import CommunitySortButtons from '@/features/community-filter/ui/CommunitySortButtons'
+import CommunitySearchForm from '@/features/community-filter/ui/CommunitySearchForm'
 
-import { createUrl } from '@/shared/lib/url'
-import SearchForm from '@/features/community-filter/ui/SearchForm'
-
-// TODO: 명세서 나오는거 보고 value 수정 & constans로 옮기기
-const CATEGORIES = [
-  { label: '전체', value: 'all' },
-  { label: '자유', value: 'free' },
-  { label: '모집', value: 'recruit' },
-  { label: '학습', value: 'study' },
-] as const
-
-const SORT = [
-  { label: '인기순', value: 'popular' },
-  { label: '최신순', value: 'latest' },
-  // { label: '오래된순', value: 'oldest' }, // 사용할 필요가..?
-] as const
-
+// TODO: searchParams 타입 정의
 interface CommunityFiltersProps {
-  activeCategory: string
-  sortBy: 'popular' | 'latest'
   searchParams: Record<string, string | string[] | undefined>
 }
 
 export default function CommunityBoardFilters({
-  activeCategory,
-  sortBy,
   searchParams,
 }: CommunityFiltersProps) {
   return (
     <div className="flex flex-col gap-4">
       {/* 윗줄 */}
       <div className="border-brand-gray-100 flex items-end justify-between border-b-2">
-        <nav className="flex gap-8">
-          {CATEGORIES.map((category) => (
-            <Link
-              key={category.value}
-              href={createUrl('', searchParams, {
-                category: category.value,
-                page: 1,
-              })}
-              className={cn(
-                'relative pb-3 text-lg font-bold transition-all',
-                activeCategory === category.value
-                  ? 'text-brand-black border-brand-black border-b-4'
-                  : 'text-brand-gray-300 hover:text-brand-gray-400'
-              )}
-            >
-              {category.label}
-            </Link>
-          ))}
-        </nav>
+        {/* 카테고리 */}
+        <CommunityCategoryTabs searchParams={searchParams} />
 
+        {/* 게시글 작성 버튼 */}
         <Link
           href={'/write'}
           className="bg-brand-black mb-2 flex items-center gap-2 rounded-lg px-6 py-3 text-base font-bold text-white hover:bg-black/80"
@@ -70,29 +35,11 @@ export default function CommunityBoardFilters({
       {/* 아랫줄 */}
       <div className="flex items-center justify-between">
         {/* 정렬 */}
-        <div className="flex shrink-0 items-center gap-2">
-          {SORT.map((type) => (
-            <Link
-              key={type.value}
-              href={createUrl('', searchParams, {
-                sort: type.value,
-                page: 1,
-              })}
-              className={cn(
-                'rounded-full border px-4 py-1.5 text-base font-bold transition-all',
-                sortBy === type.value
-                  ? 'border-brand-main text-brand-main bg-brand-main/5'
-                  : 'border-brand-gray-200 text-brand-gray-400'
-              )}
-            >
-              {type.label}
-            </Link>
-          ))}
-        </div>
+        <CommunitySortButtons searchParams={searchParams} />
 
         {/* 검색 */}
         {/* TODO: 검색 버튼 없애고 디바운스 넣거나 하려면 이젠 진짜 nuqs 쓰기 */}
-        <SearchForm currentSearchParams={searchParams} />
+        <CommunitySearchForm currentSearchParams={searchParams} />
       </div>
     </div>
   )
